LayoutTests/fast/dom/MicroData/002-expected.txt

This test ensures that document.getItems().length must return the correct number of MicroData Items in the Document. 
Also it tests that document.getItems must return a live NodeList.

document.getItems() without aurgument: PASS
document.getItems() with empty string in the aurgument: PASS
document.getItems() with 'http://example.com/foo' itemtype in the aurgument: PASS
document.getItems() with 'http://example.com/bar' itemtype in the aurgument: PASS
document.getItems() with 'http://example.com/f1' itemtype in the aurgument: PASS
Created element of type: div
Set attribute: itemscope, value: itemscope
Newly appended item should be noticed in the NodeList: PASS
Removing item should be noticed in the NodeList: PASS
PASS successfullyParsed is true

TEST COMPLETE
